backend/requirements.txt



Flask==3.0.0
Flask-Cors==4.0.0



mysql-connector-python==8.2.0
python-dotenv==1.0.0



requests==2.31.0
urllib3==2.1.0
click==8.1.7
Werkzeug==3.0.1
colorama==0.4.6